Modal de wishList adiciona e remove do Cart, porém o estado de Removido não é mantido após o fechamento do modal

+++ src/app/header-nav/wish-list/wish-list.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Renderer2, Input } from '@angular/core';
+import { Component, OnInit, Renderer2, ViewChild, ElementRef } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { WishListService } from './wish-list.service/wish-list.service';
@@ -14,8 +14,12 @@ export class WishListComponent implements OnInit {
 
   wishListSubject$: Observable<CartProduct[]>;
   
-  wishList: CartProduct[]; 
+  wishList: CartProduct[];
   
+  status = true;
+
+  @ViewChild('sBtn') sBtn: ElementRef<HTMLButtonElement>;
+
   constructor(
     private wishListService: WishListService,
     private cartService: CartService,
@@ -33,27 +37,21 @@ export class WishListComponent implements OnInit {
       this.wishListService.removeProductItem(name);
   }
 
-  addAtCart(name: string, price: number) {
+  
+  action(name:string, price:number, target:string) {
 
+    target = target.trim();
+    
+    if(target === 'Add at Cart')
       this.cartService.addProduct(name, price);
+    
+
+    if(target === 'Remove from Cart'){
+      this.cartService.removeProductItem(name);
+      this.status = false
+    }
+
   }
 
-  removeFromCart(name:string) {
-    
-    this.cartService.removeProductItem(name);
-
-  }
-
-  // actions(target:string, name: string, price: number) {
-
-  //   console.log(target === name);
-
-  //   if(target === 'Add at Cart') {
-
-  //     console.log('foi');
-  //   }else{
-  //     console.log('opa');
-  //   }
-    
-  // }
+  
 }
